+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/memory/script/MemoryScriptEngine.java
@@ -1,8 +1,8 @@
 /**
- * Copyright or © or Copr. IETR/INSA - Rennes (2014 - 2018) :
+ * Copyright or © or Copr. IETR/INSA - Rennes (2014 - 2019) :
  *
  * Alexandre Honorat <ahonorat@insa-rennes.fr> (2018)
- * Antoine Morvan <antoine.morvan@insa-rennes.fr> (2017 - 2018)
+ * Antoine Morvan <antoine.morvan@insa-rennes.fr> (2017 - 2019)
  * Clément Guy <clement.guy@insa-rennes.fr> (2014 - 2015)
  * Julien Heulot <julien.heulot@insa-rennes.fr> (2015)
  * Karol Desnos <karol.desnos@insa-rennes.fr> (2014)
@@ -214,8 +214,13 @@ public class MemoryScriptEngine {
    */
   public void generateCode(final PreesmScenario scenario, final String log) {
     final String codegenPath = scenario.getCodegenManager().getCodegenDirectory() + "/";
-
-    final IFile iFile = ResourcesPlugin.getWorkspace().getRoot().getFile(new Path(codegenPath + log + ".txt"));
+    final IFile iFile;
+    try {
+      iFile = ResourcesPlugin.getWorkspace().getRoot().getFile(new Path(codegenPath + log + ".txt"));
+    } catch (final Exception e) {
+      throw new PreesmException(
+          "Could not access code generation target path folder. Please check its value in the scenario.", e);
+    }
     try {
       ResourcesPlugin.getWorkspace().getRoot().refreshLocal(IResource.DEPTH_INFINITE, null);
       final IFolder iFolder = ResourcesPlugin.getWorkspace().getRoot().getFolder(new Path(codegenPath));